refactor: gotta go fast
The `get_currency_exchange_rate` from the `ForeignExchange` class should be written as an async method so the network IO doesn't block the event loop. This rewrite would require `get_exchange_rate` to also be async, with only the `handle_api_call` method running sync (as it is entirely CPU bound).